move server tests to correct test folder
